pyproject.toml
[project]
name = "fuzzy-broccoli"
version = "0.1.0"
description = "A production-style RESTful API demo built with FastAPI, Poetry, and Docker, showcasing clean architecture, authentication, and database integration."
authors = [
    {name = "yongli zhuang"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "black (>=25.12.0,<26.0.0)"
]

[tool.poetry]
package-mode = false
